Разбиение кода на модулю
